requirements.txt
appdirs==1.4.4
attrs==19.3.0
autopep8==1.5.3
black==19.10b0
click==7.1.2
flake8==3.8.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.3.10
passlib==1.7.2
pathspec==0.8.0
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.13.0
regex==2020.5.14
SQLAlchemy==1.3.17
toml==0.10.1
typed-ast==1.4.1
validate-email==1.3
Werkzeug==1.0.1
zipp==3.1.0
